use man2web to create html from manpage
big change cause it was not possible to split things.

create_mandb.rb now saves the output of man2web in the Berkley Database.
The controller now just adds links to the manpage and saves the output in
the cache. No extra parsing anymore

Search for manpages is now caseinsensitive, e.g. AS(1) and as(1) are the
same manpage.